# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Codex",
        "title": "API Service",
        "contact": {}
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
        "/admin/attributes/": {
            "post": {
                "description": "Adds Attribute Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Add Attribute",
                "parameters": [
                    {
                        "description": "New Attribute",
                        "name": "newAttribute",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddAttributeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Attribute Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Update Attribute",
                "parameters": [
                    {
                        "description": "Update Attribute",
                        "name": "updateAttribute",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAttributeDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/attributes/:id": {
            "delete": {
                "description": "Delete Attributes from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Delete Attribute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attribute ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/planguages/": {
            "post": {
                "description": "Adds Programming Language Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Add Programming Language",
                "parameters": [
                    {
                        "description": "New Programming Language",
                        "name": "newPLanguage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddPLanguageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates Programming Language Into DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Update Programming Language",
                "parameters": [
                    {
                        "description": "Update Programming Language",
                        "name": "updatePLanguage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePLanguageDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/planguages/{id}": {
            "delete": {
                "description": "Delete Programming Languages from DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Delete Programming Language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/private/attributes/": {
            "get": {
                "description": "Retrieves all attribute based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attributes"
                ],
                "summary": "Get All Attributes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attribute ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reward ID",
                        "name": "rewardID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "TraitType of Attribute",
                        "name": "traitType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/private/language/": {
            "get": {
                "description": "Retrieves all languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Language"
                ],
                "summary": "Get All Languages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Value",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/planguages/": {
            "get": {
                "description": "Retrieves all Programming languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Get All Programming Languages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "languageID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Programming Language Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/planguages/{id}": {
            "get": {
                "description": "Retrieves spesific Programming languages based on the provided query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Programming Language"
                ],
                "summary": "Get One Programming Language By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Programming Language ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddAttributeDTO": {
            "type": "object",
            "required": [
                "rewardID",
                "traitType",
                "value"
            ],
            "properties": {
                "rewardID": {
                    "type": "string"
                },
                "traitType": {
                    "type": "string",
                    "maxLength": 30
                },
                "value": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "dto.AddPLanguageDTO": {
            "type": "object",
            "required": [
                "fileExtention",
                "imagePath",
                "monacoEditor",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileExtention": {
                    "type": "string",
                    "maxLength": 30
                },
                "imagePath": {
                    "type": "string",
                    "maxLength": 60
                },
                "languageID": {
                    "type": "string"
                },
                "monacoEditor": {
                    "type": "string",
                    "maxLength": 30
                },
                "name": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "dto.UpdateAttributeDTO": {
            "type": "object",
            "required": [
                "rewardID"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "rewardID": {
                    "type": "string"
                },
                "traitType": {
                    "type": "string",
                    "maxLength": 30
                },
                "value": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "dto.UpdatePLanguageDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "fileExtention": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imagePath": {
                    "type": "string"
                },
                "languageID": {
                    "type": "string"
                },
                "monacoEditor": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "dataCount": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}